% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boot.lm}
\alias{boot.lm}
\title{Sampling of linear model parameters using bootstrap method.}
\usage{
boot.lm(formula, data, ...)
}
\description{
\code{boot.lm} returns the function which sample linear model parameters.
}

