% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibration.mad}
\alias{calibration.mad}
\title{calibrate zscore with mad estimator of variance}
\usage{
calibration.mad(zscore)
}
\description{
calibrate zscore with mad estimator of variance
}

