% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwas.R
\name{gwas.linearmodel.ridge}
\alias{gwas.linearmodel.ridge}
\title{X = G * B + epsilon + regularisation ridge}
\usage{
gwas.linearmodel.ridge(G, X, parameter)
}
\description{
X = G * B + epsilon + regularisation ridge
}

