% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ridge.R
\name{lfmm.ml.ridge2.lr}
\alias{lfmm.ml.ridge2.lr}
\title{LFMM with ridge likelihood implementation 2 + LR for score}
\usage{
lfmm.ml.ridge2.lr(G, X, K, lambda)
}
\description{
LFMM with ridge likelihood implementation 2 + LR for score
}

