% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ridge.R
\name{lfmm.ml.ridge2.normalization}
\alias{lfmm.ml.ridge2.normalization}
\title{LFMM with ridge likelihood implementation 2 + normalization}
\usage{
lfmm.ml.ridge2.normalization(G, X, K, lambda)
}
\description{
LFMM with ridge likelihood implementation 2 + normalization
}

