% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampler.R
\name{sample.binary.model}
\alias{sample.binary.model}
\title{Sample P(G) = logistic(mu + UV^T + XB^T + e)}
\usage{
sample.binary.model(n, L, K, sigma, prop.outlier, c, ploidy)
}
\description{
Sample P(G) = logistic(mu + UV^T + XB^T + e)
}

