% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampler.R
\name{sample.normal.model}
\alias{sample.normal.model}
\title{Sample G = mu + UV^T + XB^T + e}
\usage{
sample.normal.model(n, L, K, sigma, prop.outlier, c, mean.B = 0, sd.mu = 1,
  mean.mu = 0.5)
}
\description{
Sample G = mu + UV^T + XB^T + e
}

