% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampler.R
\name{sample.phenotype}
\alias{sample.phenotype}
\title{Sample Phenotype = Somme(B_j * G_j) + epsilon
Same simulation that in "Genomic inflation factors under polygenic inheritance"}
\usage{
sample.phenotype(G, squared.h, number.causal)
}
\description{
Sample Phenotype = Somme(B_j * G_j) + epsilon
Same simulation that in "Genomic inflation factors under polygenic inheritance"
}

